+++ portfolio/project_service/routs/project/deleteProject.js
@@ -1,6 +1,9 @@
 const Project = require('../../models/project');
 
 const deleteProject = (req, res) => {
+    if (!req.headers.authorization) {
+        return res.send({success:'user should be logged in'})
+    }
     const id = req.body.id;
     Project.findByIdAndDelete(id).then(result => {
         return res.status(200).json(result);
