+++ portfolio/skill_service/route/skills/deleteSkill.js
@@ -1,6 +1,9 @@
 const Skill = require('../../models/skill');
 
 const deleteSkill = (req, res) => {
+    if (!req.headers.authorization) {
+        return res.send({success:'user should be logged in'})
+    }
     const id = req.body.id;
     Skill.findByIdAndDelete(id).then(reslut => {
         res.send(reslut);
